Issue-38: Rename existing tests to integration
